requirements.txt
xlrd==2.0.1
openpyxl==3.0.7
pandas==1.2.4
mysql-connector-python==8.0.25